t/tree-test-flow.dat

#data
<p><article></article>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>
|     <article>

#data
<p><aside></aside>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>
|     <aside>

#skip 1
#data
<p><datagrid></datagrid>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <datagrid>

#data
<p><section></section>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>
|     <section>

#data
<p><footer></footer>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>
|     <footer>

#data
<p><header></header>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>
|     <header>

#data
<p><nav></nav>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>
|     <nav>

#data
<p><section></section>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>
|     <section>

#data
<article><span></article>x
#errors
no DOCTYPE
no </span>
#document
| <html>
|   <head>
|   <body>
|     <article>
|       <span>
|     "x"

#data
<article><p></article>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <article>
|       <p>
|     "x"

#data
<article><marquee></article>x
#errors
no DOCTYPE
unexpected end tag
no </marquee>
#document
| <html>
|   <head>
|   <body>
|     <article>
|       <marquee>
|         "x"

#data
<aside><span></aside>x
#errors
no DOCTYPE
no </span>
#document
| <html>
|   <head>
|   <body>
|     <aside>
|       <span>
|     "x"

#data
<aside><p></aside>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <aside>
|       <p>
|     "x"

#data
<aside><marquee></aside>x
#errors
no DOCTYPE
unexpected end tag
no </marquee>
#document
| <html>
|   <head>
|   <body>
|     <aside>
|       <marquee>
|         "x"

#data
<p><hgroup></hgroup>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>
|     <hgroup>

#data
<hgroup><p>aa</hgroup>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <hgroup>
|       <p>
|         "aa"

#data
<hgroup><marquee></hgroup>x
#errors
no DOCTYPE
unexpected end tag
no </marquee>
#document
| <html>
|   <head>
|   <body>
|     <hgroup>
|       <marquee>
|         "x"

#data
<datagrid><span></datagrid>x
#errors
no DOCTYPE
no </span>
#document
| <html>
|   <head>
|   <body>
|     <datagrid>
|       <span>
|     "x"

#skip 1
#data
<datagrid><p></datagrid>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <datagrid>
|       <p>
|     "x"

#data
<datagrid><marquee></datagrid>x
#errors
no DOCTYPE
unexpected end tag
no </marquee>
#document
| <html>
|   <head>
|   <body>
|     <datagrid>
|       <marquee>
|         "x"

#data
<details><span></details>x
#errors
no DOCTYPE
no </span>
#document
| <html>
|   <head>
|   <body>
|     <details>
|       <span>
|     "x"

#data
<details><p></details>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <details>
|       <p>
|     "x"

#data
<details><marquee></details>x
#errors
no DOCTYPE
unexpected end tag
no </marquee>
#document
| <html>
|   <head>
|   <body>
|     <details>
|       <marquee>
|         "x"

#data
<section><span></section>x
#errors
no DOCTYPE
no </span>
#document
| <html>
|   <head>
|   <body>
|     <section>
|       <span>
|     "x"

#data
<section><p></section>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <section>
|       <p>
|     "x"

#data
<section><marquee></section>x
#errors
no DOCTYPE
unexpected end tag
no </marquee>
#document
| <html>
|   <head>
|   <body>
|     <section>
|       <marquee>
|         "x"

#data
<figure><span></figure>x
#errors
no DOCTYPE
no </span>
#document
| <html>
|   <head>
|   <body>
|     <figure>
|       <span>
|     "x"

#data
<figure><p></figure>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <figure>
|       <p>
|     "x"

#data
<figure><marquee></figure>x
#errors
no DOCTYPE
unexpected end tag
no </marquee>
#document
| <html>
|   <head>
|   <body>
|     <figure>
|       <marquee>
|         "x"

#data
<footer><span></footer>x
#errors
no DOCTYPE
no </span>
#document
| <html>
|   <head>
|   <body>
|     <footer>
|       <span>
|     "x"

#data
<footer><p></footer>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <footer>
|       <p>
|     "x"

#data
<footer><marquee></footer>x
#errors
no DOCTYPE
unexpected end tag
no </marquee>
#document
| <html>
|   <head>
|   <body>
|     <footer>
|       <marquee>
|         "x"

#data
<header><span></header>x
#errors
no DOCTYPE
no </span>
#document
| <html>
|   <head>
|   <body>
|     <header>
|       <span>
|     "x"

#data
<header><p></header>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <header>
|       <p>
|     "x"

#data
<header><marquee></header>x
#errors
no DOCTYPE
unexpected end tag
no </marquee>
#document
| <html>
|   <head>
|   <body>
|     <header>
|       <marquee>
|         "x"

#data
<nav><span></nav>x
#errors
no DOCTYPE
no </span>
#document
| <html>
|   <head>
|   <body>
|     <nav>
|       <span>
|     "x"

#data
<nav><p></nav>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <nav>
|       <p>
|     "x"

#data
<nav><marquee></nav>x
#errors
no DOCTYPE
unexpected end tag
no </marquee>
#document
| <html>
|   <head>
|   <body>
|     <nav>
|       <marquee>
|         "x"

#data
<!DOCTYPE HTML><p><rt><address></address>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <address>

#data
<!DOCTYPE HTML><p><rt><article></article>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <article>

#data
<!DOCTYPE HTML><p><rt><aside></aside>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <aside>

#data
<!DOCTYPE HTML><p><rt><blockquote></blockquote>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <blockquote>

#data
<!DOCTYPE HTML><p><rt><center></center>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <center>

#skip 1
#data
<!DOCTYPE HTML><p><rt><datagrid></datagrid>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|         <datagrid>

#data
<!DOCTYPE HTML><p><rt><details></details>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <details>

#data
<!DOCTYPE HTML><p><rt><section></section>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <section>

#data
<!DOCTYPE HTML><p><rt><dir></dir>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <dir>

#data
<!DOCTYPE HTML><p><rt><div></div>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <div>

#data
<!DOCTYPE HTML><p><rt><dl></dl>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <dl>

#data
<!DOCTYPE HTML><p><rt><figure></figure>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <figure>

#data
<!DOCTYPE HTML><p><rt><footer></footer>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <footer>

#data
<!DOCTYPE HTML><p><rt><h4></h4>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <h4>

#data
<!DOCTYPE HTML><p><rt><header></header>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <header>

#data
<!DOCTYPE HTML><p><rt><menu></menu>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <menu>

#data
<!DOCTYPE HTML><p><rt><nav></nav>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <nav>

#data
<!DOCTYPE HTML><p><rt><ol></ol>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <ol>

#data
<!DOCTYPE HTML><p><rt><p></p>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <p>

#data
<!DOCTYPE HTML><p><rp><p></p>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rp>
|     <p>

#data
<!DOCTYPE HTML><p><option><p></p>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <option>
|     <p>

#data
<!DOCTYPE HTML><p><optgroup><p></p>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <optgroup>
|     <p>

#data
<!DOCTYPE HTML><p><rt><section></section>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <section>

#data
<!DOCTYPE HTML><p><rt><ul></ul>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <ul>

#data
<!DOCTYPE HTML><p><rt><pre></pre>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <pre>

#data
<!DOCTYPE HTML><p><rt><listing></listing>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <listing>

#data
<!DOCTYPE HTML><p><rt><form></form>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <form>

#data
<!DOCTYPE HTML><p><rt><fieldset></fieldset>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <fieldset>

#data
<!DOCTYPE HTML><p><rt><li>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <li>

#data
<!DOCTYPE HTML><p><rt><dd></dd>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <dd>

#data
<!DOCTYPE HTML><p><rt><dt>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <dt>

#data
<!DOCTYPE HTML><p><rt><plaintext>
#errors
unexpected EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <plaintext>

#data
<!DOCTYPE HTML><p><rt><table></table>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <table>

#data
<!DOCTYPE HTML><p><rt><hr>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <hr>

#data
<!DOCTYPE HTML><p><rt><isindex>
#errors
isindex
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <rt>
|     <form>
|       <hr>
|       <label>
|         "This is a searchable index. Insert your search keywords here: "
|         <input>
|           name="isindex"
|       <hr>

#data
<!DOCTYPE html><p><xmp>b</xmp>a
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|     <xmp>
|       "b"
|     "a"

#data
<!DOCTYPE html><p><x><xmp>b</xmp>a
#errors
no </x>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <x>
|     <xmp>
|       "b"
|     "a"

#data
<!DOCTYPE html><x><xmp>b</xmp>a
#errors
no </x>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <x>
|       <xmp>
|         "b"
|       "a"

#data
<!DOCTYPE html><p><a>x<xmp>b</xmp>a
#errors
no </a> before <xmp>
no </a> at the end of file
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <a>
|         "x"
|     <a>
|       <xmp>
|         "b"
|       "a"

#data
<!DOCTYPE html><p><object><xmp>a</xmp></object>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <object>
|         <xmp>
|           "a"

#data
<!DOCTYPE html><ul><li><ul></li>a
#errors
unmatched </li>
missing </ul>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <ul>
|           "a"

#data
<!DOCTYPE html><ul><li><ol></li>a
#errors
unmatched </li>
missing </ul>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <ol>
|           "a"

#data
<!DOCTYPE html><ul><li><dl></li>a
#errors
missing </dl>
missing </ul>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <dl>
|       "a"

#data
<!DOCTYPE html><ul><li><ul><li></li>a
#errors
missing </ul>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <ul>
|           <li>
|           "a"
